updated Comparison & moved to main wrapper directory
