+++ gi_wrapper_cpp/dicom_tag_strategies.h
