+++ gi_wrapper_cpp/dicom_writer.h
@@ -23,8 +23,13 @@
 
 #include <set>
 #include <string>
-#include <imebra/imebra.h>
 
+namespace gdcm {
+    class DataSet;
+    class File;
+    class PixelFormat;
+    class PhotometricInterpretation;
+}
 
 class DicomWriter
 {
@@ -33,38 +38,38 @@ public:
 
     virtual ~DicomWriter();
 
-    /// \brief Saves Gamma Index array and all necessary tags to a DICOM file.
-    ///
-    /// \note Values of parameters: dims, refXNumber, refYNumber, refZNumber, plane, refSlice
-    ///       should be the same as those passed to calculateGamma() function.
-    ///
-    /// \param filepath              Path where the output DICOM file will be saved.
-    /// \param oldDataSet            Pointer to the object in which Imebra library stored parsed reference DICOM file.
-    ///                              The pointer returned by loadDicom() function. Used while transferring tags to the output DICOM file.
-    /// \param gamma                 The calculated gamma array.
-    /// \param xNumber               Number of voxels along x axis in the gamma image.
-    /// \param yNumber               Number of voxels along y axis in the gamma image.
-    /// \param zNumber               Number of voxels along z axis in the gamma image.
-    /// \param precision             Number of digits after the decimal point in the resulting Gamma Index.
-    /// \param fillValue             Value that will be assigned to voxels for which the GI comparison was not performed.
-    /// \param rewriteTagsStrategyId Specifies the strategy of transferring tags from reference DICOM to output DICOM file:
-    ///                              - 5 - rewrite all tags, except these specified in tags parameter
-    ///                              - 4 - rewrite only these specified in tags parameter
-    ///                              - 3 - rewrite only not private tags
-    ///                              - 2 - rewrite no tags
-    ///                              - 1 or other integer values - rewrite all tags
-    /// \param tags                  (optional) Required only rewriteTagsStrategy dims is greater than 3.
-    ///                              The set specifying which exactly tags (not) to transfer.
-    ///                              For more information on how to construct this set see applyDicomTagsRewriteStrategy().
-    /// \param slice                 (optional) Required only when GI comparison was performed in 2.5D od 2dfrom3D version.
-    ///                              Specifies for which slice (in the given orientation) the GI comparison was performed.
-    ///                              First slice has number 0.
-    ///
-    ///////////////////////////////////////////////////////////////////////////////
-    void saveImage(std::string filepath, imebra::DataSet* oldDataSet, double* gamma,
-                   int xNumber, int yNumber, int zNumber, int precision, double fillValue,
-                   int rewriteTagsStrategyId, std::set<std::pair<int, int> > tags = (std::set<std::pair<int, int> >()),
-                   int slice = -1);
+/// \brief Saves Gamma Index array and all necessary tags to a DICOM file.
+///
+/// \note Values of parameters: dims, refXNumber, refYNumber, refZNumber, plane, refSlice
+///       should be the same as those passed to calculateGamma() function.
+///
+/// \param filepath              Path where the output DICOM file will be saved.
+/// \param oldFile               Reference to the object in which GDCM library stored parsed reference DICOM file.
+///                              The reference returned by loadDicom() function. Used while transferring tags to the output DICOM file.
+/// \param gamma                 The calculated gamma array.
+/// \param xNumber               Number of voxels along x axis in the gamma image.
+/// \param yNumber               Number of voxels along y axis in the gamma image.
+/// \param zNumber               Number of voxels along z axis in the gamma image.
+/// \param precision             Number of digits after the decimal point in the resulting Gamma Index.
+/// \param fillValue             Value that will be assigned to voxels for which the GI comparison was not performed.
+/// \param rewriteTagsStrategyId Specifies the strategy of transferring tags from reference DICOM to output DICOM file:
+///                              - 5 - rewrite all tags, except these specified in tags parameter
+///                              - 4 - rewrite only these specified in tags parameter
+///                              - 3 - rewrite only not private tags
+///                              - 2 - rewrite no tags
+///                              - 1 or other integer values - rewrite all tags
+/// \param tags                  (optional) Required only rewriteTagsStrategy dims is greater than 3.
+///                              The set specifying which exactly tags (not) to transfer.
+///                              For more information on how to construct this set see applyDicomTagsRewriteStrategy().
+/// \param slice                 (optional) Required only when GI comparison was performed in 2.5D od 2dfrom3D version.
+///                              Specifies for which slice (in the given orientation) the GI comparison was performed.
+///                              First slice has number 0.
+///
+///////////////////////////////////////////////////////////////////////////////
+    void saveImage(std::string filepath, gdcm::File& oldFile, double* gamma,
+        int xNumber, int yNumber, int zNumber, int precision, double fillValue,
+        int rewriteTagsStrategyId, std::set<std::pair<int, int> > tags = (std::set<std::pair<int, int> >()),
+        int slice = -1);
 
 protected:
 
@@ -73,10 +78,9 @@ protected:
     /// \note Values of parameters: xNumber, yNumber, zNumber, plane, slice
     ///       should be the same as those passed to calculateGamma() function.
     ///
-    /// \param newDataSet Object in which the new DICOM is constructed.
-    /// \param depth      The channel values data types (Imebra parameter)
-    /// \param colorSpace The Image‘s color space (Imebra parameter)
-    /// \param highBit    The highest bit occupied by the channels’ values (Imebra parameter)
+    /// \param newFile    Object in which the new DICOM is constructed.
+    /// \param pf         The pixel format of DICOM image (GDCM parameter)
+    /// \param pi         The photometric interpretation of DICOM image (GDCM parameter)
     /// \param gamma      The calculated gamma array.
     /// \param xNumber    Number of voxels along x axis in the gamma image.
     /// \param yNumber    Number of voxels along y axis in the gamma image.
@@ -89,15 +93,15 @@ protected:
     ///                   First slice has number 0.
     ///
     ///////////////////////////////////////////////////////////////////////////////
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                      double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int slice) = 0;
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int slice) = 0;
 
 private:
-    void setDicomRescaleTags(imebra::DataSet* newDataSet, int rescale);
+    void setDicomRescaleTags(gdcm::DataSet& newDataSet, int rescale);
 };
 
 
-class DicomWriter2D: public DicomWriter
+class DicomWriter2D : public DicomWriter
 {
 public:
     DicomWriter2D();
@@ -105,12 +109,11 @@ public:
     virtual ~DicomWriter2D();
 
 protected:
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                              double* gamma, int xNumber, int yNumber, int, int rescale, double fillValue, int);
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int, int rescale, double fillValue, int);
 };
 
-
-class DicomWriter3D: public DicomWriter
+class DicomWriter3D : public DicomWriter
 {
 public:
     DicomWriter3D();
@@ -118,12 +121,11 @@ public:
     virtual ~DicomWriter3D();
 
 protected:
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                              double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int);
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int);
 };
 
-
-class DicomWriter3DSliceXY: public DicomWriter
+class DicomWriter3DSliceXY : public DicomWriter
 {
 public:
     DicomWriter3DSliceXY();
@@ -131,12 +133,11 @@ public:
     virtual ~DicomWriter3DSliceXY();
 
 protected:
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                              double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int zSlice);
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int zSlice);
 };
 
-
-class DicomWriter3DSliceXZ: public DicomWriter
+class DicomWriter3DSliceXZ : public DicomWriter
 {
 public:
     DicomWriter3DSliceXZ();
@@ -144,12 +145,11 @@ public:
     virtual ~DicomWriter3DSliceXZ();
 
 protected:
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                              double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int ySlice);
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int ySlice);
 };
 
-
-class DicomWriter3DSliceYZ: public DicomWriter
+class DicomWriter3DSliceYZ : public DicomWriter
 {
 public:
     DicomWriter3DSliceYZ();
@@ -157,8 +157,8 @@ public:
     virtual ~DicomWriter3DSliceYZ();
 
 protected:
-    virtual void insertFrames(imebra::DataSet& newDataSet, imebra::bitDepth_t depth, std::string colorSpace, uint32_t highBit,
-                              double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int xSlice);
+    virtual void insertFrames(gdcm::File& newFile, gdcm::PixelFormat pf, gdcm::PhotometricInterpretation pi,
+        double* gamma, int xNumber, int yNumber, int zNumber, int rescale, double fillValue, int xSlice);
 };
 
 #endif
