+++ gi_wrapper_cpp/wrapper.cpp
@@ -20,70 +20,79 @@
 
 #include <iostream>
 #include <cstdlib>
-#include <imebra/imebra.h>
+#include "gdcmReader.h"
+#include "gdcmGlobal.h"
+#include "gdcmDicts.h"
+#include "gdcmDict.h"
+#include "gdcmAttribute.h"
+#include "gdcmStringFilter.h"
+#include "gdcmImageReader.h"
+#include "gdcmImageWriter.h"
+
 #include "dicom_reader.h"
 #include "dicom_writer.h"
 #include "wrapper_logger.h"
 
-using namespace imebra;
+using namespace gdcm;
 using namespace std;
 
 
-DataSet* loadDicom(string filepath)
+File& loadDicom(string filepath)
 {
-    logWrapperMessage("Loading DICOM file: " + filepath);
-    return CodecFactory::load(filepath);
+    Reader reader;
+    reader.SetFileName(filepath.c_str());
+    return reader.GetFile();
 }
 
-double* acquireImage(DataSet* dataSet, int& dims,
-                     double& xStart, double& xSpacing, int& xNumber,
-                     double& yStart, double& ySpacing, int& yNumber,
-                     double& zStart, double& zSpacing, int& zNumber)
+double* acquireImage(File& file, int& dims,
+    double& xStart, double& xSpacing, int& xNumber,
+    double& yStart, double& ySpacing, int& yNumber,
+    double& zStart, double& zSpacing, int& zNumber)
 {
     logWrapperMessage("Parsing image...");
     DicomReader reader;
-    return reader.acquireImage(dataSet, dims,
-                               xStart, xSpacing, xNumber,
-                               yStart, ySpacing, yNumber,
-                               zStart, zSpacing, zNumber);
+    return reader.acquireImage(file, dims,
+        xStart, xSpacing, xNumber,
+        yStart, ySpacing, yNumber,
+        zStart, zSpacing, zNumber);
 }
 
-void saveImage(int dims, DataSet* oldDataSet, string filepath, double* gamma, int refXNumber, int refYNumber, int refZNumber,
-               int precision, double fillValue, int rewriteTagsStrategy, set <pair<int, int>> tags, int plane, int refSlice)
+void saveImage(int dims, File& oldFile, string filepath, double* gamma, int refXNumber, int refYNumber, int refZNumber,
+    int precision, double fillValue, int rewriteTagsStrategy, set <pair<int, int>> tags, int plane, int refSlice)
 {
     logWrapperMessage("Creating results file at " + filepath);
 
     DicomWriter* writer;
-    switch(dims)
+    switch (dims)
     {
+    case 2:
+        writer = new DicomWriter2D();
+        break;
+    case 3:
+        writer = new DicomWriter3D();
+        break;
+    case 4:
+    case 5:
+        switch (plane)
+        {
+        case 0:
+            writer = new DicomWriter3DSliceXY();
+            break;
+        case 1:
+            writer = new DicomWriter3DSliceXZ();
+            break;
         case 2:
-            writer = new DicomWriter2D();
+            writer = new DicomWriter3DSliceYZ();
             break;
-        case 3:
-            writer = new DicomWriter3D();
-            break;
-        case 4:
-        case 5:
-            switch(plane)
-            {
-                case 0:
-                    writer = new DicomWriter3DSliceXY();
-                    break;
-                case 1:
-                    writer = new DicomWriter3DSliceXZ();
-                    break;
-                case 2:
-                    writer = new DicomWriter3DSliceYZ();
-                    break;
-                default:
-                    cerr << "Error. Wrong plane: " << plane << ". Should be in range 0-2." << endl;
-                    exit(-1);
-            }
         default:
-            cerr << "Error. Wrong dimensions number: " << dims << ". Should be in range 2-5." << endl;
+            cerr << "Error. Wrong plane: " << plane << ". Should be in range 0-2." << endl;
             exit(-1);
+        }
+    default:
+        cerr << "Error. Wrong dimensions number: " << dims << ". Should be in range 2-5." << endl;
+        exit(-1);
     }
-    writer->saveImage(filepath, oldDataSet, gamma, refXNumber, refYNumber, refZNumber, precision, fillValue, rewriteTagsStrategy, tags, refSlice);
+    writer->saveImage(filepath, oldFile, gamma, refXNumber, refYNumber, refZNumber, precision, fillValue, rewriteTagsStrategy, tags, refSlice);
     delete writer;
     logWrapperMessage("DICOM saved successfully.");
 }
