+++ gi_wrapper_cpp/wrapper_logger.h
